提交mybatis-plus自定义sql的使用

+++ SpringBootTest1/src/main/java/com/winner/mytabisplus/entity/BaseUser.java
@@ -28,7 +28,7 @@ public class BaseUser implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "BU00")
-    private Double bu00;
+    private Integer bu00;
 
     /**
      * 难搞
